New function applying syntax This is inspired by the `ap` operatior, which is better than the foldN variants.
What you do now is
targetvar <= function =$$= var1 =$= var2 =$= var3

the =$$= before the first parameter takes the function and puts it
in IO, as well starts an empty list of events to fill in by the
other =$= operators. In the end <= applies it to the function. Thanks
to Haskell, all is very nicely typed.

It's important that =< has lower precedence than =< and that they are
all left-associative.

It does not look too good when the function is an operatort, but
not worse than withN.